feat(core): refactor package structure and update product integration
