+++ src/main/java/br/matosit/inventory_service/application/dto/ProductDTO.java
@@ -1,17 +1,27 @@
 package br.matosit.inventory_service.application.dto;
 
 public class ProductDTO {
+
   private String id;
   private String name;
   private String description;
   private double price;
+  private int stockQuantity;
+  private String image3D;
 
+  public ProductDTO() {
+    // Construtor padrão
+  }
+  
   // Construtor
-  public ProductDTO(String id, String name, String description, double price) {
+  public ProductDTO(String id, String name, String description, double price, int stockQuantity,
+      String image3D) {
     this.id = id;
     this.name = name;
     this.description = description;
     this.price = price;
+    this.stockQuantity = stockQuantity;
+    this.image3D = image3D;
   }
 
   // Getters e Setters
@@ -45,5 +55,21 @@ public class ProductDTO {
 
   public void setPrice(double price) {
     this.price = price;
+  }
+
+  public int getStockQuantity() {
+    return stockQuantity;
+  }
+
+  public void setStockQuantity(int stockQuantity) {
+    this.stockQuantity = stockQuantity;
+  }
+
+  public String getImage3D() {
+    return image3D;
+  }
+
+  public void setImage3D(String image3D) {
+    this.image3D = image3D;
   }
 }
