--- src/main/java/br/matosit/inventory_service/application/ports/InventoryRepository.java
@@ -1,5 +1,0 @@
-package br.matosit.inventory_service.application.ports;
-
-public interface InventoryRepository {
-
-}
