--- src/main/java/br/matosit/inventory_service/application/ports/ProductEventHandler.java
@@ -1,7 +1,0 @@
-package br.matosit.inventory_service.application.ports;
-
-import br.matosit.inventory_service.domain.entities.Product;
-
-public interface ProductEventHandler {
-  void handleProductCreated(Product product);
-}
