--- src/main/java/br/matosit/inventory_service/application/ports/ProductRepository.java
@@ -1,9 +1,0 @@
-package br.matosit.inventory_service.application.ports;
-
-import br.matosit.inventory_service.domain.entities.Product;
-
-public interface ProductRepository {
-
-  Product save(Product customer);
-
-}
