+++ src/main/java/br/matosit/inventory_service/application/ports/in/FindInventoryUseCase.java
@@ -1,0 +1,7 @@
+package br.matosit.inventory_service.application.ports.in;
+
+import br.matosit.inventory_service.domain.entities.Product;
+
+public interface FindInventoryUseCase {
+  Product find(String productId);
+}
