+++ src/main/java/br/matosit/inventory_service/config/ApplicationProperties.java
@@ -1,0 +1,16 @@
+package br.matosit.inventory_service.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ApplicationProperties {
+
+  @Value("${webclient.product.base-url}")
+  private String webClientBaseUrl;
+
+  public String getWebClientBaseUrl() {
+    return webClientBaseUrl;
+  }
+
+}
